requirements.txt
streamlit
pandas
openpyxl
PyPDF2
python-docx
spacy
